Add VM struct
