+++ src/lib.rs
@@ -4,3 +4,4 @@ pub mod memory;
 pub mod operations;
 pub mod registers;
 pub mod utils;
+pub mod vm;
